Added proper getters and setters to NodeSpecification.scala.
Added tests, not fully functional yet.

Changed baseRequest and networkClient sendAltMessage to take in a responsemsg type, since even one way serializers actually have that type. It now passes all the tests properly, including the new test for sendAltMessage

Made a compareTo function for requestRunners that marks things with higher priority and smaller (older) times as smaller (so they come out first of a priorityQueue which defaults to min)

defaulted priority in inputSerializer to 0